chore(refactor): extract mock data for tasks to separate file
